+++ projects/common/lib/components/date/date.component.ts
@@ -250,7 +250,13 @@ You must use either [restrictDate] or the [dateRange*] inputs.
       const month = this.getNumericValue(this._month);
       const day = this.getNumericValue(this._day);
       // console.log('CREATING DATE', { year, month, day });
-      this.date = new Date(year, month, day);
+
+      // Date function appears to use setYear() so any year 0-99 results in year 1900 to 1999
+      // Set each field individually, use setFullYear() instead of setYear()
+      this.date = new Date();
+      this.date.setFullYear(year);
+      this.date.setMonth(month);
+      this.date.setDate(day);
       this._onChange(this.date);
       this.dateChange.emit(this.date);
     } else {
@@ -395,6 +401,8 @@ You must use either [restrictDate] or the [dateRange*] inputs.
   }
 
   private validateDistantDates(): ValidationErrors | null {
+
+    // console.log( 'validateDistantDates: ', distantFuture, distantPast, this.date );
 
     if (isAfter(this.date, distantFuture)) {
       return {yearDistantFuture: true};