+++ CHANGELOG.md
@@ -1,4 +1,23 @@
-# Unreleased "moh-common-lib" (2019-11-27)
+# Unreleased "moh-common-lib" (2019-12-02)
+
+
+
+# [3.0.0-rc.19](https://github.com/bcgov/moh-common-styles/compare/v3.0.0-rc.18...v3.0.0-rc.19) "moh-common-lib" (2019-12-02)
+
+
+### Bug Fixes
+
+* set each field of date separately so year does not default to 1901 when year is 1. (CR18975) ([dc606f1](https://github.com/bcgov/moh-common-styles/commit/dc606f1))
+
+
+
+# [3.0.0-rc.18](https://github.com/bcgov/moh-common-styles/compare/v3.0.0-rc.17...v3.0.0-rc.18) "moh-common-lib" (2019-12-02)
+
+
+### Features
+
+* **AbstractHttpService:** Create `uploadAttachment` method ([bf0bcc2](https://github.com/bcgov/moh-common-styles/commit/bf0bcc2))
+* **CommonImage:** Can set `fileContent` via optional constructor param ([cd1cc3b](https://github.com/bcgov/moh-common-styles/commit/cd1cc3b))